requirements.txt
redis>=5.0.0
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
requests>=2.31.0
gspread>=6.0.0
